==> Code/0_PreprocessData.ipynb <==
import numpy as np
import pandas as pd

## CICIDS-2017

# files = ['Friday-WorkingHours-Afternoon-DDos.pcap_ISCX.csv',
# 'Friday-WorkingHours-Afternoon-PortScan.pcap_ISCX.csv',
# 'Friday-WorkingHours-Afternoon-PortScan.pcap_ISCX.csv',
# 'Monday-WorkingHours.pcap_ISCX.csv',
# 'Thursday-WorkingHours-Afternoon-Infilteration.pcap_ISCX.csv',
# 'Tuesday-WorkingHours.pcap_ISCX.csv',
# 'Wednesday-workingHours.pcap_ISCX.csv']

# d = pd.DataFrame()
# for f in files:
#     raw = pd.read_csv('../Data/CICIDS_2017/' + f)
#     d = pd.concat([d, raw], axis=0)
d = pd.read_csv('../Data/CICIDS_2017/all.csv')
print(d.shape)
d = d.dropna()
print(d.shape)

train = pd.DataFrame()
for i in range(5):
    train = pd.concat([train, d.sample(100000)], axis=0)
train.to_csv('../Data/CICIDS_2017/train.csv')
print(train.shape)
test= pd.DataFrame()
for i in range(5):
    test = pd.concat([test, d.sample(100000)], axis=0)
test.to_csv('../Data/CICIDS_2017/test.csv')
print(test.shape)

## NSL-KDD

from scipy.io.arff import loadarff

data = loadarff('../Data/NSL_KDD/KDDTrain+.arff')
df = pd.DataFrame(data[0])
cat = df.select_dtypes(include='O').columns
df[cat] = df[cat].apply(lambda x: x.str.decode('utf8'))
df.to_csv('../Data/NSL_KDD/train.csv', index=False)

data = loadarff('../Data/NSL_KDD/KDDTest+.arff')
df = pd.DataFrame(data[0])
cat = df.select_dtypes(include='O').columns
df[cat] = df[cat].apply(lambda x: x.str.decode('utf8'))
df.to_csv('../Data/NSL_KDD/test.csv', index=False)
